Enable multi-language support on site
This sets up a folder structure and configuration to enable future work
to add multiple languages to the site. To add another language, create
content under an approprite content/ subdirectory and add the key to the
'languages' param in config. As soon as a second language is added, a
dropdown menu will be generated in the Docsy top nav to select the
language.