Implement the Docsy theme (#561)
* Enable Docsy theme

* Set config params for use with Docsy theme

* Remove layout files which directly conflict with Docsy theme

* Add overriding blog content template

This template overrides the default Docsy file - it accounts for
multiple authors, as well as authors containing attributes. By default,
in Docsy the post's 'author' parameter is a string containing the
author's name.

* Update docs subsection layout with Docsy markup

* Add custom sidebar-nav with icons for external links

* Add hooks partials as per Docsy guidelines

Docsy provides `head-end.html` and `body-end.html` as entry-points to
extending these areas of the theme. Existing partials for applicable
scripts and stylesheets have been migrated to these new partials.

* Add theme style variables override file

* Remove Algolia DocSearch styles and script

* Enable Google search with placeholder engine

* Add optimized header logo to Docsy header

* Add shortcode for building 'Used By' logos

* Implement home page

Refactors hompage to make the switch over from Bulma to Bootstrap.
This also entails reworking some of the shortcodes, and using Docsy's
built-in features wherever possible, such as the blocks/cover shortcode.

* Disable alternate layout for root language pages in docs

* Render auth_alts pages with shortcode instead of custom layout

* Use CSS to set background on cover blocks

* Implement 'about' page

Adds content file and shortcodes as needed to implement the 'About' page
using the Docsy theme using Docsy shortcodes.

* Implement 'showcase' page

Adds content file and shortcodes as needed to implement the 'Showcase'
page using the Docsy theme using Docsy shortcodes.

* Implement 'community' page

Adds content file and shortcodes as needed to implement the 'Community'
page using the Docsy theme using Docsy shortcodes.

* Override Docsy's list community layout

Docsy has a default 'list.html' layout file, which is opinionated about
the opening copy and block structure of this page.

* Import new site stylesheet into overridden main imports

* Apply external link icon styles

* Size homepage logo responsively

* Hide root /docs leaf pages from sidebar nav

* Move FAQ to the /docs/what_is_grpc section

* Add new in-page TOC for large screens and down

This overrides the default Docsy 'baseof' for the docs section, as well
as a new partial and styles to wrap the default table of contents in an
expandable menu.

* Clean up docs index pages for basic list layout

* Implement 'alert' Docsy shortcode in content

Docsy has a single shortcode for handling the type of functionality that
several shortcodes are currently. Whether the Docsy 'alert' shortcode is
further customized later or not, this change will allow several files to
be replaced with one.

* Replace Reddit with Google Group link in footer

* Override list page layout for the docs section

This new override is in place to account for custom frontmatter params.
Currently, the index pages for individual programming languages are
given a custom layout.

* Replace alt layout with flag for language indexes

This replaces the use of an alternate layout file for programming
language home pages with a flag that can lean on the newly overidden
'docs/list.html' template file.

* Add comments to site-specific stylesheet

* Disable Docsy overrides of default Bootstrap buttons

* Add new shortcode for rendering ToC on main pages

* Override Docsy secondary color with 'gRPC blue'

* Hide subsection list on Overview page

* Override Hugo sitemap and hide legacy pages on it

The 'quickstart' and 'tutorials' pages under the root /docs folder are
to be left, but hidden from in-page navigation and indexing. This uses
the example-site Docsy sitemap.xml file to override the default Hugo
one. A frontmatter param of 'private' is used as a flag to hide the page
from the sitemap.

* Remove unused layout, partials and shortcodes

* Disable ToC on language list pages

* Translate image rendering from Bulma to Bootstrap

* Add bannner partial to main page templates

This refactors the optional banner set in the site config from Bulma to
a custom class. In order to include the partial in the base site
structure, the default 'baseof' layout file has been overriden.

* Override breadcrumb partial to display on mobile as well

* Override Docsy CSS partial with diffs from old CSS partial

* Extract language home page content to partial

* Refactor 404 page to use Docsy-like structure and classes

* Translate buttons shortcode from Bulma to Bootstrap

* Refactor contribute page to have same structure as the other standalone pages

* Fix overflow issue with docs content on small mobile screens

* Fix user logo height on 'about' page

* Apply base and responsive styling to homepage images

* Support short titles for side nav and breadcrumbs

This change makes a check for a frontmatter param for shortened titles
while building links. It also revises the name of the param from the
existing "short" to "short_title".

* Enable "simple" view of languages index page

Frontmatter param of "simple_list" is provided by Docsy

* Remove javascript partial no longer required

As Alpine is no longer in use to trigger modals and the anchor tag
offset has been generated by Docsy, the JavaScript partial can be
removed.

* Revert auto-format changes

* Give same max-width to main content as container

In the docs section and the blog, the max-width is defined by Docsy as
80%. This intentionally caps that value at the same max-width defined by
the 'container' used elsewhere for very large screens.

* Remove list bullets and increase spacing on prog-lang-home pages

* Add redirect rule for FAQ page

* Specify 'main' branch when building meta links

The gRPC repo has updated its "master" branch to the gaining-acceptance
"main" naming.

* Enable multi-language support on site

This sets up a folder structure and configuration to enable future work
to add multiple languages to the site. To add another language, create
content under an approprite content/ subdirectory and add the key to the
'languages' param in config. As soon as a second language is added, a
dropdown menu will be generated in the Docsy top nav to select the
language.

* Revert git command in netlfy build steps

* Give supported language pages custom titles

Since this is overriding Docsy's "head" partial, the ranging over the
default "alternative output formats" and creating broken links has been
commented out of the partial as well.

* Revert "Revert git command in netlfy build steps"

This reverts commit 60a6fc076fb4a7f141f092e016c8d3cb21044532.

* Disable Docsy buttons using variable override

* Remove deprecation warning on use of '.Dir'