Pass over styles to implement a namespacing scheme
This revises much of the inline styles using Bootstrap utilities to try
and limit their use to Bootstrap components and concerns with display
and layout. Some styling has been extracted to namespaced classes when
either the object in question had potential for reuse, extracting
improved clarity, or the content was encapsulated on a top-level page.

An exception has been made to the inline table of contents on top-pages,
as this has a dependency on Docsy's internal ToC styling, as well as our
overrides for use in the /docs section. Its inline styling has been
minimized.