Comment files which are overriding Docsy or Hugo
Overrides incur a degree of technical debt as they will need to be kept
up to date when the Docsy theme is updated, if there are any conflicts
or new functionality which needs to be migrated. If the need for an
override or piece of customization disappears, the file can be safely
deleted as Hugo which default to Docsy in its lookup order.