Refactor Docs ToC and improve styles (#567)
* Refactor inline ToC approach and add second link

This revises the approach to expanding and collapsing the inline ToC in
the docs pages - rather than template login, a more CSS-driven solution
was adopted. A fix to the shortcode ToC and some improved comments on
the styles have also been added.

* Optimize common background SVG

* Revise style structure and class names on pages

* Move Gitter to social links

* Render images with Markdown parser where able

* Fill view height with docsy ToC

* Fix reference to font awesome icon

* Pass over styles to implement a namespacing scheme

This revises much of the inline styles using Bootstrap utilities to try
and limit their use to Bootstrap components and concerns with display
and layout. Some styling has been extracted to namespaced classes when
either the object in question had potential for reuse, extracting
improved clarity, or the content was encapsulated on a top-level page.

An exception has been made to the inline table of contents on top-pages,
as this has a dependency on Docsy's internal ToC styling, as well as our
overrides for use in the /docs section. Its inline styling has been
minimized.