Restore Docsy version to commit @f7f3c42
A unintentional regression was introduced in PR#561 to the Docsy
submodule. This points the submodule to the original intended commit,
and performs maintenance on customizations to the Docsy theme,
incorporating theme updates to files which are currently overridden.

For the full diff of changes to Docsy in this update, see:
https://github.com/google/docsy/compare/053672...f7fec42