Use absolute paths instead of absolute URLs
When you specify grpc.io in a URL, it is considered a separate site and
gets the "linking away" anchor behavior, with the icon and opening in a
new tab. Also, such links don't work properly when running locally or
using the PR deploy preview.

In metadata.md, the concepts guide moved locations (see
layouts/index.redirects), so let's inline that redirect.